fix: migration files, foreign keys, seeders connected shipments page to the backend, added approve and reject buttons to documents

+++ database/migrations/2026_04_26_131417_create_role_permissions_table.php
@@ -9,8 +9,8 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('role_permissions', function (Blueprint $table) {
-            $table->unsignedInteger('role_id');
-            $table->unsignedInteger('permission_id');
+            $table->unsignedBigInteger('role_id');
+            $table->unsignedBigInteger('permission_id');
 
             $table->primary(['role_id', 'permission_id']);
